requirements.txt
asgiref==3.11.0
Django==5.2.10
django-environ==0.12.0
django-tenants==3.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
